refactor: modularize API clients
