+++ src/services/api/index.ts
@@ -1,0 +1,16 @@
+export * from './artist';
+export * from './admin';
+export * from './auth';
+export * from './base';
+export * from './category';
+export * from './community';
+export * from './constants';
+export * from './events';
+export * from './funding';
+export * from './gallery';
+export * from './health';
+export * from './interaction';
+export * from './liveStream';
+export * from './stats';
+export * from './user';
+export * from './userProfile';
